+++ cli.ts
@@ -20,16 +20,13 @@ denopack.option(
   {}
 );
 
-denopack
-  .option("--watch <dirOrFile>", "Watch a file or directory and rebuild on changes", {})
-  .usage("--watch mod.ts")
-  .usage("--watch src");
+denopack.option("--cache <cacheLocation>", "Persist build cache", {});
 
-denopack.option("--cache <cacheLocation>", "Persist build cache", {}).usage("--cache cache.json");
-
-denopack.option("-p, --print", "Prints the generated bundle to stdout", {});
+denopack.option("--watch <dirOrFile>", "Watch a file or directory and rebuild on changes", {});
 
 denopack.option("--defaultConfig", "Prints the default config to stdout", {});
+
+denopack.option("--print", "Prints the generated bundle to stdout", {});
 
 denopack.help(() => {
   console.log("🦕📦🦕📦🦕📦🦕📦\n");
@@ -39,7 +36,7 @@ denopack.example("denopack -i mod.ts");
 denopack.example("denopack -i mod.ts -o bundle.js");
 denopack.example("denopack -i mod.ts --dir dist");
 denopack.example("denopack -c denopack.config.ts");
-denopack.example("denopack -i mod.ts -o out.js --dir dist -c denopack.config.ts");
+denopack.example("denopack -i mod.ts -o out.js -d dist -c denopack.config.ts");
 
 const opts = denopack.parse().options as Options;